third_party/WebKit/LayoutTests/crypto/subtle/wrapKey-unextractable-expected.txt

Tests that an unextractable key cannot be wrapped.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS typeof key.extractable is 'boolean'
PASS key.extractable is false
error is: InvalidAccessError: key is not extractable
PASS successfullyParsed is true

TEST COMPLETE
